Fixed a broken migration
Requires manual dropping and creation of a function to retroactively fix in production.

+++ migrations/20170331095601_wilsons_as_function.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
   return knex.schema.raw(`
       CREATE FUNCTION wilsons(ups int, downs int) RETURNS numeric LANGUAGE plpgsql IMMUTABLE AS $$
       DECLARE
-        z int;
+        z numeric;
         n int;
         p numeric;
         l numeric;
